Java version changed to 11. Vector values changed to double.

+++ lib/src/main/java/edu/austral/dissis/starships/vector/Vector2.java
@@ -5,27 +5,27 @@ import org.jetbrains.annotations.NotNull;
 import static java.lang.Math.*;
 
 public class Vector2 {
-    private final float x;
-    private final float y;
+    private final double x;
+    private final double y;
 
-    private Vector2(float x, float y) {
+    private Vector2(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getAngle() {
-        return (float) (atan2(y, x) - atan2(0, 1));
+    public double getAngle() {
+        return atan2(y, x) - atan2(0, 1);
     }
 
-    public float getModule() {
-        return (float) Math.pow(Math.pow(x, 2) + Math.pow(y, 2), 0.5);
+    public double getModule() {
+        return Math.pow(Math.pow(x, 2) + Math.pow(y, 2), 0.5);
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
@@ -40,28 +40,28 @@ public class Vector2 {
     }
 
     @NotNull
-    public Vector2 multiply(Float scalar) {
+    public Vector2 multiply(double scalar) {
         return new Vector2(x * scalar, y * scalar);
     }
 
     @NotNull
-    public Vector2 rotate(float angle) {
-        return new Vector2((float) (x * cos(angle) - y * sin(angle)), (float) (x * sin(angle) + y * cos(angle)));
+    public Vector2 rotate(double angle) {
+        return new Vector2((double) (x * cos(angle) - y * sin(angle)), (double) (x * sin(angle) + y * cos(angle)));
     }
 
     @NotNull
     public Vector2 asUnitary() {
-        final float module = getModule();
+        final double module = getModule();
         return new Vector2(x / module, y / module);
     }
 
     @NotNull
-    public static Vector2 vector(float x, float y) {
+    public static Vector2 vector(double x, double y) {
         return new Vector2(x, y);
     }
 
     @NotNull
-    public static Vector2 vectorFromModule(float module, float angle) {
-        return new Vector2((float) (module * cos(angle)), (float) (module * sin(angle)));
+    public static Vector2 vectorFromModule(double module, double angle) {
+        return new Vector2((double) (module * cos(angle)), (double) (module * sin(angle)));
     }
 }
